refactor: split the four files still over 1,200 lines
Each split is by concept, not by line count. Where the two disagreed I moved
the boundary, and two of them did disagree - noted below, because a split cut
on adjacency makes a file shorter and the code no better.

## `components.rs` 1,306 → 373 + 951

`impl ContextWindow` was 838 lines of the file on its own. It is now
`components/context_window.rs`: the window, its assembly into a request, and its
eviction. What remains in `components/` is what an agent *is* - its state,
status, wait reason, parentage, inbox. "What an agent is" and "what an agent
remembers" are different questions to arrive with.

## `lint.rs` 1,209 → 374 + 519 + 342

`lint/checks.rs` reads a manifest as a shape: does every stage it names exist,
can the run reach an output, does a tool it advertises resolve. `lint/security.rs`
reads the same manifest as a request for capability: tool policies, command
seeds, pre-approved commands, read-path grants, held checkpoints. Those answer
"will this agent work" and "should this agent be allowed to", which is a real
line rather than a convenient one. `mod.rs` keeps the vocabulary and the
dispatcher.

## `host/mod.rs` 1,454 → 566 + four sections

`health` (lane pressure and what `lev daemon status` reports), `emit` (the
change-detection pass that decides what is worth broadcasting), `subagents`
(operations that arrive from inside the world rather than from a client), and
`listing` (what `lev ps` sees).

**Two boundaries were wrong on the first pass and are fixed here.** Retention
(`record_finished` / `prune_finished`) landed in `health` because it sat next to
the lane code in the original file - but it is bookkeeping on the finished-run
*list*, so it belongs with `listing`, where `finished()` already was. And
`subagent_sender` landed in `emit` because it sat next to `subscribe`; it hands
out the channel sub-agent ops arrive on, so it belongs with `subagents`. Both
moved, and the module docs now describe what each file actually holds.

## `setup/state.rs` 1,241 → 793 + 250 + 169

`types` is the vocabulary the renderer and the input handler both speak; `limits`
is the `[limits]` screen and the two suggested write ceilings a fresh install
gets. The wizard itself stays in `mod.rs`.

**Same correction here:** `effort_options` / `effort_index` were in `limits`
because of where they sat. They describe a provider row's reasoning effort and
have nothing to do with `[limits]`, so they moved beside `ProviderRow` in
`types`. `limits.rs` now holds only limits.

## Verification

Every split is verified pure by sorting all non-comment, non-import lines before
and after: `components` 3,402 → 3,402 with **zero** difference, `lint` 872 → 872,
and the rest differing only by `pub(super)` on methods the move forced across a
module boundary and one `impl WorldHost {` per new section.

6,891 workspace tests pass; `leviath-runtime` and `leviath-cli` both hold 100%.

Files over 1,200 production lines: **4 → 1** (`daemon/spawn.rs`, left alone
because #311 is already editing it).
