+++ crates/leviath-cli/src/commands/setup/state/mod.rs
@@ -17,229 +17,12 @@ use super::verify::Outcome;
 use crate::bundled::{AgentAction, BundledAgent};
 use crate::config::Config;
 
-/// The wizard's screens, in order.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-pub enum Step {
-    /// What the wizard is about to do, before it touches anything.
-    Welcome,
-    /// Pick which providers to configure.
-    Providers,
-    /// Enter or confirm a credential for each picked provider.
-    ProviderDetail,
-    /// The default provider and model new runs use.
-    Defaults,
-    /// Concurrency, timeouts, and the other numeric ceilings.
-    Limits,
-    /// Choose which bundled blueprints to install.
-    Agents,
-    /// Import MCP servers found in other harnesses' configs.
-    Mcp,
-    /// The whole plan, before anything is written.
-    Review,
-}
-
-impl Step {
-    /// Every step, in order.
-    pub const ALL: [Step; 8] = [
-        Step::Welcome,
-        Step::Providers,
-        Step::ProviderDetail,
-        Step::Defaults,
-        Step::Limits,
-        Step::Agents,
-        Step::Mcp,
-        Step::Review,
-    ];
-
-    /// Title shown in the header.
-    pub fn title(self) -> &'static str {
-        match self {
-            Step::Welcome => "Welcome",
-            Step::Providers => "Providers",
-            Step::ProviderDetail => "Credentials",
-            Step::Defaults => "Defaults",
-            Step::Limits => "Limits",
-            Step::Agents => "Agents",
-            Step::Mcp => "MCP servers",
-            Step::Review => "Review",
-        }
-    }
-
-    /// Position in [`Self::ALL`].
-    pub fn index(self) -> usize {
-        Self::ALL
-            .iter()
-            .position(|s| *s == self)
-            .expect("every step is in ALL")
-    }
-}
-
-/// One row of the provider pick-list.
-#[derive(Debug, Clone)]
-pub struct ProviderRow {
-    /// Which provider this row is for.
-    pub provider: Provider,
-    /// Whether the user has picked it.
-    pub selected: bool,
-    /// The credential as typed. Empty means "no value".
-    pub value: String,
-    /// The credential is already in the environment, under this variable, and
-    /// is not being written to the config.
-    pub from_env: Option<&'static str>,
-    /// Reasoning effort, for the Claude Code transport only.
-    pub effort: usize,
-    /// What the last verification attempt concluded.
-    pub outcome: Outcome,
-    /// A verification is in flight.
-    pub checking: bool,
-}
-
-impl ProviderRow {
-    /// Whether this provider has something to verify.
-    pub fn has_credential(&self) -> bool {
-        match self.provider.credential {
-            Credential::ApiKey => !self.value.is_empty() || self.from_env.is_some(),
-            // Ollama and the Claude Code transport need no key; selecting them
-            // is the whole configuration, so they are always checkable.
-            Credential::BaseUrl | Credential::None => true,
-        }
-    }
-}
-
-/// One row of the blueprint list.
-#[derive(Debug, Clone)]
-pub struct AgentRow {
-    /// The bundled blueprint this row offers.
-    pub agent: &'static BundledAgent,
-    /// What installing it would do: a fresh install, an upgrade, or nothing
-    /// because the same version is already there.
-    pub action: AgentAction,
-    /// Whether the user has picked it.
-    pub selected: bool,
-}
-
-/// One importable MCP server.
-#[derive(Debug, Clone)]
-pub struct McpRow {
-    /// The server definition as found, before collision handling.
-    pub candidate: Candidate,
-    /// Which harness it came from.
-    pub source: String,
-    /// Whether the user has picked it.
-    pub selected: bool,
-    /// A server of this name is already in the Leviath config.
-    pub collides: bool,
-    /// The name it will actually be stored under, after collision handling.
-    pub name: String,
-}
-
-/// A single editable setting on the Defaults / Limits screens.
-#[derive(Debug, Clone)]
-pub struct Field {
-    /// The setting's name, as shown.
-    pub label: &'static str,
-    /// One line explaining what it does, shown under the label.
-    pub help: &'static str,
-    /// Its current value, which also decides what a key press means.
-    pub value: FieldValue,
-}
-
-/// The kinds of setting the wizard edits, and therefore the ways a key press
-/// can mean something.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum FieldValue {
-    /// A whole number. `None` means unset.
-    Number(Option<u64>),
-    /// A toggle.
-    Bool(bool),
-    /// One of a fixed list.
-    Choice {
-        /// Every value this field may take, in the order they are cycled.
-        options: Vec<String>,
-        /// Which one is selected, by position in `options`.
-        index: usize,
-    },
-}
-
-impl FieldValue {
-    /// How the value reads on screen.
-    pub fn display(&self) -> String {
-        match self {
-            Self::Number(None) => "(unset)".to_string(),
-            Self::Number(Some(n)) => n.to_string(),
-            Self::Bool(true) => "yes".to_string(),
-            Self::Bool(false) => "no".to_string(),
-            Self::Choice { options, index } => match options.get(*index) {
-                Some(chosen) => chosen.clone(),
-                None => "(none)".to_string(),
-            },
-        }
-    }
-
-    /// The options of a choice field; empty for any other kind.
-    pub fn options(&self) -> &[String] {
-        match self {
-            Self::Choice { options, .. } => options,
-            Self::Number(_) | Self::Bool(_) => &[],
-        }
-    }
-}
-
-/// Asked of the background verifier.
-#[derive(Debug, Clone)]
-pub struct VerifyRequest {
-    /// Which provider to check, matching the row that asked.
-    pub provider_id: String,
-    /// The credentials to check, as typed or taken from the environment.
-    pub creds: leviath_runtime::provider_creds::ProviderCreds,
-}
-
-/// Answered by the background verifier.
-#[derive(Debug, Clone)]
-pub struct VerifyReply {
-    /// Which provider this answers for. Replies can arrive out of order, so
-    /// this is what routes one back to its row.
-    pub provider_id: String,
-    /// What the check concluded.
-    pub outcome: Outcome,
-}
-
-/// Where the text being typed goes when it is committed.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum EditTarget {
-    /// The credential of the provider at this index in `providers`.
-    Credential(usize),
-    /// The field at this index of the current step's fields.
-    Field(usize),
-}
-
-/// An in-progress text edit.
-#[derive(Debug, Clone)]
-pub struct Edit {
-    /// Where the text goes when it is committed.
-    pub target: EditTarget,
-    /// The shared single-line editor (cursor movement, masking).
-    pub(crate) line: crate::tui::widgets::line_edit::LineEdit,
-}
-
-/// Why a confirmation dialog is on screen, so its Yes can be routed.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ConfirmPurpose {
-    /// `q`/Ctrl-C with unsaved choices: quit and discard?
-    QuitDiscard,
-    /// Saving with the Claude Code transport selected: accept the terms risk?
-    SaveTos,
-    /// Leaving the Providers screen with nothing selected: continue anyway?
-    NoProviders,
-}
-
-/// A pending confirmation: the dialog plus what its Yes means.
-#[derive(Debug, Clone, PartialEq)]
-pub struct PendingConfirm {
-    /// What answering Yes would mean.
-    pub purpose: ConfirmPurpose,
-    pub(crate) dialog: crate::tui::widgets::confirm::Confirm,
-}
+// Sections of the former single-file wizard state. Glob re-exported so every
+// existing `state::Wizard` path keeps working.
+mod limits;
+use limits::*;
+mod types;
+pub(crate) use types::*;
 
 /// The whole wizard.
 pub struct Wizard {
@@ -1031,180 +814,6 @@ impl Wizard {
             .filter(|r| r.selected && !r.candidate.inline_secrets.is_empty())
             .map(|r| format!("{}: {}", r.name, r.candidate.inline_secrets.join(", ")))
             .collect()
-    }
-}
-
-/// Reasoning-effort levels for the Claude Code transport, from the provider
-/// rather than re-typed here.
-pub fn effort_options() -> &'static [&'static str] {
-    &leviath_providers::claude_code::EFFORT_LEVELS
-}
-
-/// Index of `effort` in [`effort_options`], defaulting to the provider default.
-fn effort_index(effort: Option<&str>) -> usize {
-    let wanted = effort.unwrap_or(leviath_providers::claude_code::DEFAULT_EFFORT);
-    effort_options()
-        .iter()
-        .position(|e| *e == wanted)
-        .unwrap_or_default()
-}
-
-/// The Limits screen's fields, seeded from a config.
-fn limits_fields(config: &Config) -> Vec<Field> {
-    vec![
-        Field {
-            label: "Max concurrent inferences",
-            help: "How many model calls run at once across all agents.",
-            value: FieldValue::Number(config.limits.max_concurrent_inferences.map(|n| n as u64)),
-        },
-        Field {
-            label: "Max concurrent tools",
-            help: "How many tool calls run at once within one batch.",
-            value: FieldValue::Number(Some(config.limits.max_concurrent_tools as u64)),
-        },
-        Field {
-            label: "Default max iterations",
-            help: "Per-stage ceiling when a blueprint sets none.",
-            value: FieldValue::Number(config.limits.default_max_iterations.map(|n| n as u64)),
-        },
-        Field {
-            label: "Exact token counting",
-            help: "Ask the provider for real token counts instead of estimating. Slower.",
-            value: FieldValue::Bool(config.limits.exact_token_counting),
-        },
-        Field {
-            label: "Batch tool-call hint",
-            help: "Nudge models to request several tools in one turn.",
-            value: FieldValue::Bool(config.batch_tool_hint),
-        },
-        Field {
-            label: "Platform shell hint",
-            help: "Tell models what shell they get. Only says anything on Windows (cmd.exe).",
-            value: FieldValue::Bool(config.shell_hint),
-        },
-        Field {
-            label: "Stall timeout (seconds)",
-            help: "Fail a run that can never dispatch (unconfigured provider). 0 waits forever.",
-            value: FieldValue::Number(Some(config.limits.stall_timeout_secs)),
-        },
-        Field {
-            label: "Dead cycles before relief",
-            help: "Widen the tool lane after this many 30s cycles with work queued and nothing moving. 0 never does.",
-            value: FieldValue::Number(Some(config.limits.dead_cycles_before_relief as u64)),
-        },
-        Field {
-            label: "Finished run retention (seconds)",
-            help: "Keep a run in `lev ps` this long after it ends, so a script polling on an interval sees how it ended. 0 drops it at once.",
-            value: FieldValue::Number(Some(config.limits.finished_retention_secs)),
-        },
-        Field {
-            label: "Wedge timeout (seconds)",
-            help: "Fail a run nothing in the engine can reach any more. 0 is off; 300 is a sensible value.",
-            value: FieldValue::Number(Some(config.limits.wedge_timeout_secs)),
-        },
-        Field {
-            label: "Interaction timeout (seconds)",
-            help: "Resolve a prompt nobody answered after this long, so the run carries on. 0 waits for ever.",
-            value: FieldValue::Number(Some(config.limits.interaction_timeout_secs)),
-        },
-        // The two write ceilings are unset in code and offered with a value
-        // here, so a fresh install has one written down where it can be seen
-        // and cleared. Clearing the field stores nothing, which is unlimited.
-        Field {
-            label: "Max bytes one tool call may write",
-            help: "Stops a single command that writes until the disk fills. Clear it for no limit.",
-            value: FieldValue::Number(Some(
-                config
-                    .limits
-                    .max_tool_call_write_bytes
-                    .unwrap_or(SUGGESTED_CALL_WRITE_BYTES),
-            )),
-        },
-        Field {
-            label: "Max bytes one run may write",
-            help: "The same ceiling across a whole run, which several ordinary-looking calls can reach. Clear it for no limit.",
-            value: FieldValue::Number(Some(
-                config
-                    .limits
-                    .max_run_write_bytes
-                    .unwrap_or(SUGGESTED_RUN_WRITE_BYTES),
-            )),
-        },
-    ]
-}
-
-/// What `lev setup` offers as a per-call write ceiling.
-///
-/// 2 GiB is far above anything a tool call legitimately writes - a large build
-/// log, a full database dump - and far below the 14 GB a single runaway append
-/// reached in the incident this exists for. The gap between those two numbers
-/// is wide enough that the value does not need to be right, only present.
-const SUGGESTED_CALL_WRITE_BYTES: u64 = 2 * 1024 * 1024 * 1024;
-
-/// What `lev setup` offers as a per-run write ceiling.
-///
-/// Five times the per-call figure, so a run doing genuinely heavy work has room
-/// for several large writes while three runaway calls still stop.
-const SUGGESTED_RUN_WRITE_BYTES: u64 = 5 * SUGGESTED_CALL_WRITE_BYTES;
-
-/// Write the Limits screen's fields back into a config.
-fn apply_limits_fields(config: &mut Config, fields: &[Field]) {
-    for (index, field) in fields.iter().enumerate() {
-        match (index, &field.value) {
-            (0, FieldValue::Number(n)) => {
-                config.limits.max_concurrent_inferences = n.map(|n| n as usize)
-            }
-            // A zero here would deadlock every tool batch, so an explicit unset
-            // or 0 falls back to the default rather than being stored.
-            (1, FieldValue::Number(n)) => {
-                config.limits.max_concurrent_tools = n
-                    .filter(|n| *n > 0)
-                    .map(|n| n as usize)
-                    .unwrap_or(Config::default().limits.max_concurrent_tools)
-            }
-            (2, FieldValue::Number(n)) => {
-                config.limits.default_max_iterations = n.map(|n| n as usize)
-            }
-            (3, FieldValue::Bool(b)) => config.limits.exact_token_counting = *b,
-            (4, FieldValue::Bool(b)) => config.batch_tool_hint = *b,
-            (5, FieldValue::Bool(b)) => config.shell_hint = *b,
-            // Unset means "leave the watchdog at its default", not "disable it";
-            // disabling is an explicit 0.
-            (6, FieldValue::Number(n)) => {
-                config.limits.stall_timeout_secs =
-                    n.unwrap_or(Config::default().limits.stall_timeout_secs)
-            }
-            // Same rule: unset keeps the default, 0 is an explicit "never".
-            (7, FieldValue::Number(n)) => {
-                config.limits.dead_cycles_before_relief = n
-                    .map(|n| n as u32)
-                    .unwrap_or(Config::default().limits.dead_cycles_before_relief)
-            }
-            // And again: unset keeps the default, 0 means keep nothing.
-            (8, FieldValue::Number(n)) => {
-                config.limits.finished_retention_secs =
-                    n.unwrap_or(Config::default().limits.finished_retention_secs)
-            }
-            // Same rule once more, and here the default is itself 0 (off).
-            (9, FieldValue::Number(n)) => {
-                config.limits.wedge_timeout_secs =
-                    n.unwrap_or(Config::default().limits.wedge_timeout_secs)
-            }
-            // Same rule again: unset keeps the default hour, 0 is an explicit
-            // "wait for a person however long it takes".
-            (10, FieldValue::Number(n)) => {
-                config.limits.interaction_timeout_secs =
-                    n.unwrap_or(Config::default().limits.interaction_timeout_secs)
-            }
-            // The write ceilings break the rule above, and deliberately: here
-            // an unset field means *no limit*, not "keep the default". They are
-            // the only two settings whose absence is a real choice a user makes
-            // - deleting the line is how you say "let it write" - so unset
-            // stores `None` rather than reinstating a number they just removed.
-            (11, FieldValue::Number(n)) => config.limits.max_tool_call_write_bytes = *n,
-            (12, FieldValue::Number(n)) => config.limits.max_run_write_bytes = *n,
-            _ => {}
-        }
     }
 }
