+++ crates/leviath-runtime/src/host/subagents.rs
@@ -1,0 +1,222 @@
+//! Sub-agent operations, which are the only control ops an *agent* can issue.
+//!
+//! Spawning a child, checking on one, cancelling a subtree. Kept apart from the
+//! control loop because these arrive from inside the world (an agent's tool
+//! call) rather than from a client, and the tree walks they need - ancestry,
+//! cancellation - exist nowhere else. The channel they arrive on is handed out
+//! here too, so the sender and its only reader are in one file.
+
+use super::*;
+
+impl WorldHost {
+    /// Service one [`SubAgentOp`] from a tool lane, replying on its oneshot.
+    pub(super) fn handle_subagent(&mut self, op: SubAgentOp) {
+        match op {
+            SubAgentOp::Spawn {
+                args,
+                parent_run_id,
+                max_depth,
+                reply,
+            } => {
+                let _ = reply.send(self.spawn_child(*args, &parent_run_id, max_depth));
+            }
+            SubAgentOp::Check { run_id, reply } => {
+                let report = self.live_entity(&run_id).and_then(|e| {
+                    self.world.agent_status(e).map(|status| SubAgentReport {
+                        status,
+                        final_output: self
+                            .world
+                            .world()
+                            .get::<crate::persistence::FinalOutput>(e)
+                            .map(|o| o.0.clone()),
+                    })
+                });
+                let _ = reply.send(report);
+            }
+            SubAgentOp::Send {
+                run_id,
+                caller_run_id,
+                content,
+                target_region,
+                reply,
+            } => {
+                if !self.is_within_tree(&run_id, &caller_run_id) {
+                    let _ = reply.send(false);
+                    return;
+                }
+                // Page the target in if it was unloaded, so delivery finds it.
+                self.resolve_or_reload(&run_id);
+                let ok = self
+                    .world
+                    .send_message(AgentMessage {
+                        agent_id: run_id,
+                        content,
+                        target_region,
+                    })
+                    .is_ok();
+                let _ = reply.send(ok);
+            }
+            SubAgentOp::Kill {
+                run_id,
+                caller_run_id,
+                reply,
+            } => {
+                let within = self.is_within_tree(&run_id, &caller_run_id);
+                let _ = reply.send(within && self.cancel_tree(&run_id));
+            }
+        }
+    }
+
+    /// Spawn a child agent under `parent_run_id`, linking `ParentRef` /
+    /// `SubAgentChildren` and registering its run id. `Err` if the parent is not
+    /// live, the depth limit is reached, or the spawner rejects it.
+    pub(super) fn spawn_child(
+        &mut self,
+        mut args: SpawnArgs,
+        parent_run_id: &str,
+        max_depth: usize,
+    ) -> Result<String, String> {
+        // Record the parentage so the child's run metadata nests it in the tree.
+        args.parent_run_id = Some(parent_run_id.to_string());
+        let parent = self
+            .live_entity(parent_run_id)
+            .ok_or_else(|| format!("parent run '{parent_run_id}' is not live"))?;
+        let parent_depth = self
+            .world
+            .world()
+            .get::<ParentRef>(parent)
+            .map_or(0, |p| p.depth);
+        let child_depth = parent_depth + 1;
+        if child_depth > max_depth {
+            return Err(format!(
+                "sub-agent depth limit ({max_depth}) reached; not spawning deeper"
+            ));
+        }
+        let run_id = args.run_id.clone();
+        let child = match self.spawner.as_mut() {
+            Some(spawner) => spawner(&mut self.world, &args)?,
+            None => return Err("this daemon cannot spawn agents".to_string()),
+        };
+        let world = self.world.world_mut();
+        world.entity_mut(child).insert(ParentRef {
+            parent_entity: parent,
+            parent_agent_id: parent_run_id.to_string(),
+            depth: child_depth,
+        });
+        match world.get_mut::<SubAgentChildren>(parent) {
+            Some(mut kids) => kids.children.push(child),
+            None => {
+                world.entity_mut(parent).insert(SubAgentChildren {
+                    children: vec![child],
+                    max_child_depth: max_depth,
+                });
+            }
+        }
+        // Record the child's run-id on the parent's serializable state so the
+        // tree is persisted (and restart can rebuild `SubAgentChildren`). A
+        // spawning parent always carries `AgentState`.
+        world
+            .get_mut::<crate::components::AgentState>(parent)
+            .expect("a spawning parent always has AgentState")
+            .spawned_children_ids
+            .push(run_id.clone());
+        // Seed the child's context from the parent per any declared blueprint
+        // context transform (planner→coder region mapping, etc.).
+        crate::context_transform::apply_context_transforms(world, parent, child);
+        self.by_run_id.insert(run_id.clone(), child);
+        Ok(run_id)
+    }
+
+    /// Cancel a run and every descendant, paging the root in from disk first if it
+    /// had been unloaded. Returns whether the run was found in the world.
+    ///
+    /// Cancelling only the root would leave its sub-agents and fan-out workers
+    /// running - they are independent agents the schedule keeps driving, so they
+    /// would carry on spending tokens with no parent to report to. Each cancelled
+    /// agent's open interactions are closed too, so nothing is left blocked on a
+    /// prompt for a run that is going away.
+    /// Whether `run_id` is `ancestor` itself or one of its descendants.
+    ///
+    /// `send_to_agent` and `kill_agent` took any run id at all. Nothing tied the
+    /// target to the caller, so an agent could cancel an unrelated run, inject
+    /// text into its context, or - worst - hand it data: a message is added to
+    /// the target as `Public` regardless of the sender's taint, so an agent
+    /// holding `Private` context whose own outbound tools were gated could pass
+    /// it to a sibling whose tools were not. That is a laundering channel
+    /// straight through the middle of taint tracking.
+    ///
+    /// A downward walk from the caller, the same shape [`cancel_tree`] uses:
+    /// parentage is recorded as `SubAgentChildren`, so "is it mine" is "is it in
+    /// my subtree".
+    ///
+    /// [`cancel_tree`]: Self::cancel_tree
+    pub(super) fn is_within_tree(&mut self, run_id: &str, ancestor: &str) -> bool {
+        if run_id == ancestor {
+            return true;
+        }
+        // Both ends as entities: the host already maps run ids to them, and
+        // comparing entities avoids re-reading an id component per node.
+        let (Some(target), Some(root)) = (
+            self.resolve_or_reload(run_id),
+            self.resolve_or_reload(ancestor),
+        ) else {
+            return false;
+        };
+        let mut stack = vec![root];
+        while let Some(e) = stack.pop() {
+            if e == target {
+                return true;
+            }
+            if let Some(kids) = self.world.world().get::<SubAgentChildren>(e) {
+                stack.extend(kids.children.iter().copied());
+            }
+        }
+        false
+    }
+
+    pub(super) fn cancel_tree(&mut self, run_id: &str) -> bool {
+        let Some(root) = self.resolve_or_reload(run_id) else {
+            return false;
+        };
+        // Collect the subtree (parent before children), then cancel each.
+        let mut subtree = Vec::new();
+        let mut stack = vec![root];
+        while let Some(e) = stack.pop() {
+            subtree.push(e);
+            if let Some(kids) = self.world.world().get::<SubAgentChildren>(e) {
+                stack.extend(kids.children.iter().copied());
+            }
+        }
+        let mut cancelled = false;
+        for e in subtree {
+            // Read the agent id before cancelling - the entity stays valid until
+            // it is reaped, but reading first keeps this independent of that.
+            let agent_id = self
+                .world
+                .world()
+                .get::<AgentState>(e)
+                .map(|s| s.agent_id.clone());
+            cancelled |= self.world.cancel(e);
+            if let Some(agent_id) = agent_id {
+                self.interactions.cancel_for_agent(&agent_id);
+                // The hub is keyed by agent id but the emitted-interaction set is
+                // keyed by request id, so drop the ids that are no longer pending.
+                let still_open: HashSet<String> = self
+                    .interactions
+                    .pending()
+                    .into_iter()
+                    .map(|(_, req)| req.id)
+                    .collect();
+                self.emitted_interactions
+                    .retain(|id| still_open.contains(id));
+            }
+        }
+        cancelled
+    }
+
+    /// A sender for [`SubAgentOp`]s. The daemon hands a clone to each agent's tool
+    /// state so the sub-agent tools can reach the world through the host.
+    pub fn subagent_sender(&self) -> UnboundedSender<SubAgentOp> {
+        self.subagent_tx.clone()
+    }
+}
